Trait modalcell::Behavior

pub unsafe trait Behavior {
    type Container<T>: Clone + Deref<Target = T>;

    // Required methods
    fn new<T>(value: T) -> Self::Container<T>;
    fn ptr_eq<T>(a: &Self::Container<T>, b: &Self::Container<T>) -> bool;
}

A trait controlling how SharedMode and all associated types store reference-counted values.

Safety

When ptr_eq is provided two Containers from separate calls to new(), it always returns false. When ptr_eq is provided the same Container in both a and b, it always returns true.

Breaking this guarantee is undefined behavior.

Required Associated Types

type Container<T>: Clone + Deref<Target = T>

The container type used by this tag type.

Required Methods

fn new<T>(value: T) -> Self::Container<T>

Returns a new container wrapping value.

fn ptr_eq<T>(a: &Self::Container<T>, b: &Self::Container<T>) -> bool

Returns true if a and b represent the same underlying memory allocation.